[#8] Move TestHelper to test folder.
