+++ clients/adapter-kafka/src/test/java/org/eclipse/hono/adapter/client/telemetry/kafka/KafkaBasedTelemetrySenderTest.java
@@ -24,7 +24,6 @@ import org.apache.kafka.common.header.internals.RecordHeader;
 import org.eclipse.hono.config.KafkaProducerConfigProperties;
 import org.eclipse.hono.kafka.client.CachingKafkaProducerFactory;
 import org.eclipse.hono.kafka.client.HonoTopic;
-import org.eclipse.hono.kafka.client.test.TestHelper;
 import org.eclipse.hono.util.QoS;
 import org.eclipse.hono.util.RegistrationAssertion;
 import org.eclipse.hono.util.TenantObject;
